IDP_Project/requirements.txt
opencv-python==4.5.5.64
pytesseract==0.3.13
pandas==1.3.3
pdf2image==1.16.3
openpyxl==3.0.9
Pillow
numpy

